Updates to resolution switching and shutter gating.
Add the minimum frame time to the frameGeometry class to try and
pass the expected frame timing to the sensor's setResolution()
method. This allows operating at reduced resolutions with the full
wavetable length for best quality.

We also make a first-pass attempt at adding the exposureMode
parameter and the selection of the shutterGating timing program.