Add Conflicts dependency between cam-control and camApp
For the time being, the control API and the C++ GUI are both trying
to initialize and control the FPGA; and this often leaves the FPGA
in a broken state during init. Until camApp has been refactored
accordingly, ensure that only one of the two ever runs at any given
time by adding the Conflicts= directive to the service files.