Rename IO for 'cpuInt' to 'trigger', fixed segment bugs.
The recording end trigger that exists in camApp got mixed up with the CPU
interrupt signal in the new IO block. Since the CPU interrupt isn't
actually hooked up yet, we rename it to 'trigger' since that's closer
to what we are likely to use it for in the meantime.

There were also a few minor bugs in the recording sequencer program for
segmented mode that needed some cleanup.