Update debian scripts for automated package builds.
The changelog generation script may need to operate from an anonymous
git checkout, in which case the username and e-mail will not be set.
We can get them instead from the most recent commit to the repository.

We also add a helper script conditionally install packages into a
reprepro package repository only if the version number has changed.