lux2100: Fixup ADC calibration errors.
Some lingering code from the old 3-point calibration was failing when the
midpoint was too close to the lowpoint, but this doesn't matter anymore
because the midpoint isn't actually used, so we can just remove it.

The ADC offset training also needs to take into account the weird channel
mapping between SCI register address and the logical ADC channel number
to which it applies.