Refactor power properties, add battery critical tracking.
This attempts to fix a number of issues with the power class in that it
was unable to report changes, could not query the configuration from the
PMIC, and had many unfinished parameters revolving around the auto-powerdown
mode.

The API changes include:
  - powerDownOnLowBattery properties and friends have been removed.
  - Add batteryCritical to report when a shutdown is imminent due to battery charge.
  - Add fanOverride parameter to disable the fan (boolean).
  - shippingMode can now be set via the API.

During the process, we also refactored the property handling so that
the camera class can inherit properties from its member classes. Hopefully
this will allow us to split the camera class up into logical subclasses.

Closes: https://github.com/krontech/pychronos/issues/51
Closes: https://github.com/krontech/pychronos/issues/47
Closes: https://github.com/krontech/pychronos/issues/41
Closes: https://github.com/krontech/pychronos/issues/28