Set wbTemperature to 0K aftter auto white balance.
After computing a custom white balance, we must set the wbTemperature
to zero to select it, otherwise the WB-by-temperature will override
the custom white balance we just computed.

This is part of a fix for camApp not being able to set a custom white
balance.

See: https://github.com/krontech/chronos-cam-app/issues/114