timing: Adjust register map based on version.
I have a work in progress with the FPGA bitstream to add some extra
registers to the timing program block, but to do so the programming
address needs to move. To keep pychronos compatible with the FPGA
development builds, we need to check the version to determine the
register layout.