Added import and export functions for lux2100 column gain cal.
Implemented import and export functions for column gain .bin
files that will be processed off-camera. Modified loadAnalogCal
to attempt to load the new/factory processed cal files, falling
back on the 2-point auto cal if necessary. Replaced references
to regGainBit with regSerialGain, as the former seemed to be
carried over from the 1310 driver.

TODO: automate a run of black cal prior to starting export for
better frames, use the serial number as the folder name for
exporting frames.