Fix reboot logic when reverting settings.
When calling the reboot API with just a revert-to-default-settings
option specified, the control API would exit and never restart. Instead
we should make the 'reload' option default to true unless expliclty
set to false to ensure the API reboots itself.

While we're at it, the 'defaults' option isn't well named and I feel
that change it to 'settings' would be more clear about what is being
reset when this option is provided.