Change fanOverride to a float, to set speed directly.
This gives the user more precise control over the desired fan speed
rather than just setting it on or off. fanOverride now takes a float
in the range of 0.0 (off) to 1.0 (full speed) to specify the desired
speed, or any negative value to set the fan to automatic mode based
on temperature.

Closes: https://github.com/krontech/pychronos/issues/56