lux2100: Fix inconsistent wavetable selection.
There were a bunch of different wavetable selection functions in the
LUX2100 driver, each of which were slightly different. This would
result in unexpected changes between wavetables, and invalid wavetable
choices. This should fix a bug where selecting the 'max' framerate
would lead to corrupted video.

It's worth noting that unlike the lux1310, it seems that the lux2100
requires the length of the wavetable to be greater than or equal to
the line period.