Remove calibration on resolution change, send calSuggested instead.
This feature turned out to break some factory calibration methods by
putting the camera into a busy state when the factory calibration was
about to start. Instead we should merely 'suggest' that the GUI perform
calibration if there is no calibration data present.

To that end, we have added a new parameter to the api: calSuggested
which toggles on and off whenever the camera is unable to load suitable
calibration data.

See: https://github.com/krontech/chronos-cam-app/issues/150