lux2100: Fix rounding bug causing bad wavetable selection.
When parsing a frame period into an integer number of clock periods
a rouding error would wind up choosing the wavetable for the desired
frame period minus one clock. This would tend to result in the 'max'
button in the GUI tending to select the next-shortest wavetable
often resulting in sub-standard image quality or missing cal data.