Remove fanOverride from settings.
We think that this option is a bit dangerous, and can get the user into
a state with a non-functional camera due to thermal overload. Therefore
we would like to remove it from the saved settings and return back to
auto on each boot.

See: https://github.com/krontech/chronos-cam-app/issues/146