lux2100: Set line delay to 15 clocks for all wavetables.
Just setting the line delay to 15 clocks appears to work best for all
the wavetables that we are currently able to test, and except for WT25
this appears to fix the horizontal line bug.

For WT25, the line delay appears to make no difference to the horizontal
shutter line. So we just choose 15 clocks in this case for consistency.