Fix typos and timing bugs in sensor timing programs.
There was a particularly nasty typo in the timing programs where the
run bit wasn't actually being set or cleared when swapping programs
which would lead to deadlocks and race conditions in the sensor.

Once we fixed that, we also discovered that the lux2100 had a race
condition where the frame geometry could be updated while the wavetable
was still active, leading to a deadlocked image sensor, so we add some
more care to the updating of the readout window.