Workaround race condition when reading live frames.
Since the GPMC memory bus is only 16-bits wide, it's not possible to
atomically read a 32-bit address out of the FPGA. When the FPGA happens
to update the address in the middle of a read, we can get a mangled
frame address. This tends to manifest as bizarre or invalid frame data
during calibration at small frame sizes.

See: https://github.com/krontech/chronos-cam-app/issues/154