lux1310: Retry the reset if chip detection fails.
We have seen a small probability of cameras that fail to detect their
image sensors after a reset. The problem manifests as an unexpected
value of 0x48 appearing instead of the normal 0xda. This issue seems
to go away by simply resetting the chip again. This is probably being
caused by the weird inferred clock getting out of sync between the
FPGA and the LUX1310.