Add API call to configure video playback modes.
At present the only way to switch the camera into playback mode is to
use the playback() method in the video system, but this isn't actually
exposed anywhere in the control API, and hence it isn't possible to
control playback remotely.

This adds a startPlayback() method to the control API that acts as a
proxy to the Video system.

See: https://github.com/krontech/pychronos/issues/60