Add Calls for video state changes, parse parameter docstrings.
There are a bunch of small bits in this commit, the first is to add
API calls to manipulate the video state (live, playback, and abort
of recording), and also to remove the softReset call.

We were also in the process of adding docstring parsing for API
parameters, so that got rolled in too.