Fix rollover in auto white balance.
When white balance is attmepted with the aperture or shutter closed,
the corrected pixel value due to FPN can be negative, causing the
white balance algorithm to think that the image is actually clipping.

Unfortunately, numpy seems to be of no help here, so we instead need
to manually check for pixel rollovers when calculating the average
channel values.

Closes: https://github.com/krontech/chronos-cam-app/issues/161