externalStorage: Remove disk usage information.
When parsing the disk usage information, the underlying statfs() call can
take a long time, or even hang the system, when network storage is being
used. This can slow down the system to the point of unusability, so to
work around the problem we will just remove the usage statistics, and
gather the storage information from the mount information.

Closes: https://github.com/krontech/pychronos/issues/39