Occasional check to sync with FPGA sequencer state
This makes it so that when the IO block triggers a 'start recording', the rest of the camera reacts accordingly -- GUI changes to record mode, API reports state change, red LEDs illuminate on camera, etc. The caveat is that it only checks the status at 1 second intervals, so there can be up to a 1 second delay between the signal and the camera's reaction (although the footage is still properly timed with the signal).